优化
1、学校抢课系统分为三大类，原代码无法抢本学期必修及专业内跨年级，测试修复了此BUG。
2、删除了原修改源码的使用方式，增加了交互语句。
3、打包成.exe文件，以后无需安装python环境以及依赖包。
